Refactor(App): Remove unused state and components for cleaner code
